New simulation runs

+++ simulations/_get_configs_from_easysquid.py
@@ -14,8 +14,9 @@ import easysquid
 
 path_to_this_file = os.path.realpath(__file__)
 path_to_this_folder = "/".join(path_to_this_file.split("/")[:-1])
-path_to_this_config_folder = os.path.join(path_to_this_folder, "create_measure_simulation/setupsim/config")
-other_folders = [os.path.join(path_to_this_folder, conf_folder) for conf_folder in ["major_simulation/setupsim/config"]]
+path_to_one_config_folder = os.path.join(path_to_this_folder, "create_measure_simulation/setupsim/config")
+other_folders = ["major_simulation", "many_simulations", "request_freq_sweep", "fidelity_sweep"]
+path_to_other_folders = [os.path.join(path_to_this_folder, folder, "setupsim", "config") for folder in other_folders]
 
 NODE_CENTRIC_HERALDED_FIBRE_CONNECTION = "node_centric_heralded_fibre_connection"
 
@@ -23,12 +24,13 @@ NODE_CENTRIC_HERALDED_FIBRE_CONNECTION = "node_centric_heralded_fibre_connection
 def _remove_current_files(path):
     for folder in os.listdir(path):
         dst = os.path.join(path, folder)
-        if os.path.exists(dst):
-            shutil.rmtree(dst)
+        if os.path.isdir(dst):
+            if os.path.exists(dst):
+                shutil.rmtree(dst)
 
 
 def copy_files_from_easysquid():
-    _remove_current_files(path_to_this_config_folder)
+    _remove_current_files(path_to_one_config_folder)
 
     path_to_easysquid___init__ = os.path.abspath(easysquid.__file__)
     path_to_easysquid = "/".join(path_to_easysquid___init__.split("/")[:-2])
@@ -36,7 +38,7 @@ def copy_files_from_easysquid():
     path_to_network_configs = os.path.join(path_to_easysquid, "config/networks/NV")
     for folder in os.listdir(path_to_network_configs):
         src = os.path.join(path_to_network_configs, folder)
-        dst = os.path.join(path_to_this_config_folder, folder)
+        dst = os.path.join(path_to_one_config_folder, folder)
         shutil.copytree(src, dst)
 
 
@@ -48,60 +50,62 @@ def copy_qlink_wc_wc_high_loss():
     qlink_path = "qlink/networks_with_cavity_with_conversion.json"
     qlink_high_c_loss_path = "qlink/networks_with_cavity_with_conversion_high_c_loss.json"
     easysquid_qlink_wc_wc_path = os.path.join(path_to_network_configs, qlink_path)
-    qlinklayer_qlink_wc_wc_path = os.path.join(path_to_this_config_folder, qlink_high_c_loss_path)
+    qlinklayer_qlink_wc_wc_path = os.path.join(path_to_one_config_folder, qlink_high_c_loss_path)
     shutil.copy(easysquid_qlink_wc_wc_path, qlinklayer_qlink_wc_wc_path)
 
 
 def change_connnection_type():
-    for dirpath, dirname, filenames in os.walk(path_to_this_config_folder):
+    for dirpath, dirname, filenames in os.walk(path_to_one_config_folder):
         for filename in filenames:
-            file_path = os.path.join(dirpath, filename)
+            if filename.endswith(".json"):
+                file_path = os.path.join(dirpath, filename)
 
-            # Read config file
-            with open(file_path, 'r') as f:
-                config_dct = json.load(f)
+                # Read config file
+                with open(file_path, 'r') as f:
+                    config_dct = json.load(f)
 
-            # Set electron T2 to zero
-            for q in config_dct["qpd_config"]["default"]["parameters"]["qubits"]:
-                if q["qubit_type"] == "electron":
-                    q["noise_model"]["T1T2"]["T2"] = 1000000.0
+                # Set electron T2 to zero
+                for q in config_dct["qpd_config"]["default"]["parameters"]["qubits"]:
+                    if q["qubit_type"] == "electron":
+                        q["noise_model"]["T1T2"]["T2"] = 1000000.0
 
-            # Get the connection config name of the mhp connection
-            conn_config_name = _get_conn_config_name_of_mhp_conn(config_dct)
+                # Get the connection config name of the mhp connection
+                conn_config_name = _get_conn_config_name_of_mhp_conn(config_dct)
 
-            # Update the type to use node centric heralded fibre connection
-            config_dct["conn_configs"][conn_config_name]["type"] = NODE_CENTRIC_HERALDED_FIBRE_CONNECTION
+                # Update the type to use node centric heralded fibre connection
+                config_dct["conn_configs"][conn_config_name]["type"] = NODE_CENTRIC_HERALDED_FIBRE_CONNECTION
 
-            # Add virtual delay to shorter length fibre to  make sure messages delivered in same mhp cycle
-            cycle_period = config_dct["conn_configs"][conn_config_name]["parameters"].get("t_cycle", None)
-            if cycle_period is None:
-                device_config = config_dct["qpd_config"]["default"]
-                photon_time = device_config["parameters"]["photon_emission"]["photon_emission_delay"]
-                meas_time = device_config["parameters"]["gates"]["electron_gates"]["measurement_op"]["operation_time"]
-                time_window = config_dct["conn_configs"][conn_config_name]["parameters"]["time_window"]
-                cycle_period = max(photon_time + meas_time, time_window)
-                cycle_period += cycle_period / 10
+                # Add virtual delay to shorter length fibre to  make sure messages delivered in same mhp cycle
+                cycle_period = config_dct["conn_configs"][conn_config_name]["parameters"].get("t_cycle", None)
+                if cycle_period is None:
+                    device_config = config_dct["qpd_config"]["default"]
+                    photon_time = device_config["parameters"]["photon_emission"]["photon_emission_delay"]
+                    electron_gates = device_config["parameters"]["gates"]["electron_gates"]
+                    meas_time = electron_gates["measurement_op"]["operation_time"]
+                    time_window = config_dct["conn_configs"][conn_config_name]["parameters"]["time_window"]
+                    cycle_period = max(photon_time + meas_time, time_window)
+                    cycle_period += cycle_period / 10
 
-            lengthA = config_dct["conn_configs"][conn_config_name]["parameters"]["lengthA"]
-            lengthB = config_dct["conn_configs"][conn_config_name]["parameters"]["lengthB"]
-            c = config_dct["conn_configs"][conn_config_name]["parameters"]["c"]
-            delayA = 1e9 * lengthA / c
-            delayB = 1e9 * lengthB / c
-            delay_max, delay_min = max(delayA, delayB), min(delayA, delayB)
-            virtual_delay = (2 * delay_max - delay_min)
-            delay_spec = [delay_min, virtual_delay]
-            if lengthA < lengthB:
-                config_dct["conn_configs"][conn_config_name]["parameters"]["delay_A"] = delay_spec
-            elif lengthB < lengthA:
-                config_dct["conn_configs"][conn_config_name]["parameters"]["delay_B"] = delay_spec
+                lengthA = config_dct["conn_configs"][conn_config_name]["parameters"]["lengthA"]
+                lengthB = config_dct["conn_configs"][conn_config_name]["parameters"]["lengthB"]
+                c = config_dct["conn_configs"][conn_config_name]["parameters"]["c"]
+                delayA = 1e9 * lengthA / c
+                delayB = 1e9 * lengthB / c
+                delay_max, delay_min = max(delayA, delayB), min(delayA, delayB)
+                virtual_delay = (2 * delay_max - delay_min)
+                delay_spec = [delay_min, virtual_delay]
+                if lengthA < lengthB:
+                    config_dct["conn_configs"][conn_config_name]["parameters"]["delay_A"] = delay_spec
+                elif lengthB < lengthA:
+                    config_dct["conn_configs"][conn_config_name]["parameters"]["delay_B"] = delay_spec
 
-            # Update the comment in the file
-            config_dct["AutoGenerate"].append("This file was then later modified by /path/to/QLinkLayer/simulations/"
-                                              "_get_configs_from_easysquid.py")
+                # Update the comment in the file
+                config_dct["AutoGenerate"].append("This file was then later modified by /path/to/QLinkLayer/"
+                                                  "simulations/_get_configs_from_easysquid.py")
 
-            # Write to file again
-            with open(file_path, 'w') as f:
-                json.dump(config_dct, f, indent=2)
+                # Write to file again
+                with open(file_path, 'w') as f:
+                    json.dump(config_dct, f, indent=2)
 
 
 def _get_conn_config_name_of_mhp_conn(config_dct):
@@ -127,11 +131,11 @@ def _get_qpd_config_name_of_qpd(config_dct):
 
 def make_no_loss_and_no_noise_files():
     # Create folder
-    no_noise_folder = os.path.join(path_to_this_config_folder, "no_noise")
+    no_noise_folder = os.path.join(path_to_one_config_folder, "no_noise")
     os.mkdir(no_noise_folder)
 
     # Create no losses file
-    src = os.path.join(path_to_this_config_folder, "lab", "networks_no_cavity_no_conversion.json")
+    src = os.path.join(path_to_one_config_folder, "lab", "networks_no_cavity_no_conversion.json")
     dst = os.path.join(no_noise_folder, "no_losses.json")
     shutil.copyfile(src, dst)
     _update_no_losses_file(dst)
@@ -215,8 +219,8 @@ def add_loss_qlink_wc_wc(p_c_loss):
     elif isinstance(p_c_loss, float):
         qlink_wc_wc_path = "qlink/networks_with_cavity_with_conversion.json"
         qlink_high_c_loss_path = "qlink/networks_with_cavity_with_conversion_high_c_loss_{:.0e}.json".format(p_c_loss)
-        qlinklayer_qlink_wc_wc_path = os.path.join(path_to_this_config_folder, qlink_wc_wc_path)
-        qlinklayer_qlink_wc_wc_high_c_loss_path = os.path.join(path_to_this_config_folder, qlink_high_c_loss_path)
+        qlinklayer_qlink_wc_wc_path = os.path.join(path_to_one_config_folder, qlink_wc_wc_path)
+        qlinklayer_qlink_wc_wc_high_c_loss_path = os.path.join(path_to_one_config_folder, qlink_high_c_loss_path)
 
         # Read config file
         with open(qlinklayer_qlink_wc_wc_path, 'r') as f:
@@ -241,12 +245,13 @@ def add_loss_qlink_wc_wc(p_c_loss):
 
 
 def copy_files_to_other_folders():
-    for other_folder in other_folders:
+    for other_folder in path_to_other_folders:
         _remove_current_files(other_folder)
-        for folder in os.listdir(path_to_this_config_folder):
-            src = os.path.join(path_to_this_config_folder, folder)
-            dst = os.path.join(other_folder, folder)
-            shutil.copytree(src, dst)
+        for folder in os.listdir(path_to_one_config_folder):
+            src = os.path.join(path_to_one_config_folder, folder)
+            if os.path.isdir(src):
+                dst = os.path.join(other_folder, folder)
+                shutil.copytree(src, dst)
 
 
 def main():
